fgffdfd

+++ src/main/java/com/B2Boost/RestAPIProblem/Service/PartnerService.java
@@ -1,10 +1,13 @@
 package com.B2Boost.RestAPIProblem.Service;
 
+import com.B2Boost.RestAPIProblem.Exception.BadRequestException;
+import com.B2Boost.RestAPIProblem.Exception.ResourceNotFoundException;
 import com.B2Boost.RestAPIProblem.Model.Partner;
 import com.B2Boost.RestAPIProblem.Repository.PartnerRepository;
 import jakarta.transaction.Transactional;
 import org.hibernate.query.Page;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -20,12 +23,22 @@ public class PartnerService {
     @Autowired
     public PartnerRepository repo;
 
-    public List<Partner> getAllPartners(){
-        Pageable pages = PageRequest.of(0, 10, Sort.Direction.ASC, "id");
+    public List<Partner> getAllPartners(int from, int size)  {
+        Pageable pages = PageRequest.of(from, size, Sort.Direction.ASC, "id");
 
+
+        //return repo.findAll(pages).getContent();
+        List<Partner> partners = repo.findAll(pages).getContent();
+//        if(partners.isEmpty()){
+//            throw new BadRequestException();
+//        } else {
+//            return partners;
+//        }
         return repo.findAll(pages).getContent();
     }
     public Partner savePartner(Partner partner){
+
+
         return repo.save(partner);
     }
     public Partner getPartnerById(Long id){
@@ -36,22 +49,29 @@ public class PartnerService {
 //        } else {
 //            throw new RuntimeException("Doesn't find by id ");
 //        }
-        return repo.findById(id).orElse(null);
+        return repo.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-//    public void DeletePartnerById(Long id){
-//         repo.deleteById(id);
-//    }
 
     public void deletePartnerById(Long id) {
+
+        if(repo.existsById(id)){
             repo.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException(id);
+        }
+        //Ici on peut utiliser Optional partner = repo.findById et ensuite throw si y a une exception
     }
-//        } else {
-//            return false; // Partner not found
-//        }}
 
     public Partner UpdatePartner(Partner partner){
-        return repo.save(partner);
+        if(repo.existsById(partner.getId())){
+            return repo.save(partner);
+        } else {
+            throw new ResourceNotFoundException(partner.getId());
+        }
+
+        //Pour le 404, je ne sais pas si c'est logique qu'on fasse le 404 car si il existe pas il sera créé non ?
     }
 
 }
